Create add fonts to python
Needed to download fonts as will be attempting for the NN to learn reasoning of numbers with a dataset of numbers in different fonts. The premise is that the NN will understand the basic form of letters and number and understand that there are similarity in numbers across different fronts but the object  and value remains the same across all variations. we are doing this as worry about memorization and over-fitting with one dataset in the same font. we have 8 fonts to use.